Created Edit Button and Input for Focus Timer
I created the edit button on the activity cards but couldn't figure out the functionality of actually editing the title of the activities. Functions like the delete button for now. The input for the length of focus timer has been created but the storing of that input data needs to be coded. I will continue to work on that.